fix request_id_default_length
Actually the standard library function token_urlsafe(n) returns a string
longer than n bytes/characters, because it is a base64 representation of
a random byte string of length n. But we will trim the token to be exactly
n characters long so there are no weird surprises around the meaning of
request_id_default_length.